+++ src/types/DocType.ts
@@ -1,0 +1,9 @@
+import DocField from './DocField';
+
+type DocType = {
+  name: string;
+  description: string | null;
+  fields: DocField[] | null;
+};
+
+export default DocType;
